// projects/obs_overlay_current_music_sources_linux/frontend/src/helper/types.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type ArrayOf_MediaSource = MediaSource[];

export interface MediaSource {
  bus_name: string;
  desktop_entry?: string | null;
  identity?: string | null;
  position?: Position | null;
  status?: string | null;
  track?: Track | null;
  [k: string]: unknown;
}
export interface Position {
  current_us?: number | null;
  length_us?: number | null;
  [k: string]: unknown;
}
export interface Track {
  album?: string | null;
  album_artists: string[];
  art_url?: string | null;
  artists: string[];
  genres: string[];
  length_us?: number | null;
  title?: string | null;
  track_id?: string | null;
  track_number?: number | null;
  url?: string | null;
  [k: string]: unknown;
}